Day 14 -- batch processing elements on demand
Process only part of the input at a time so we don't keep the whole thing
in memory. Zoom in on the space between the first pair for a bunch of
iterations, then save that and zoom in on the space between the first pair
in the result. Repeat until the desired level of iteration and then get the
next pair from the next shallower partial result.
(so excited this actually works!)